Initial Commit - Enemy Aggro and Vision
The short goal of this one is that I want to create an enemy that reacts to seeing the player, but only if that player is unobscured by an object. This has been done countless times in video games but I figure why not give it a go on my own in Godot.